+++ packaging/chocolatey/tools/chocolateyInstall.ps1
@@ -1,9 +1,9 @@
 $ErrorActionPreference = 'Stop'
 
 $packageName  = 'localpdf-studio'
-$version      = '0.17.2'
+$version      = '0.18.0'
 $url64        = "https://github.com/jeiel85/localpdf-studio/releases/download/v$version/LocalPDF.Studio_${version}_x64-setup.exe"
-$checksum64   = 'bb4bedb4ad3d39e0407202defe1301600a4901034f3ad96320c7b1ac95ea7d81'
+$checksum64   = '62cf45f8be320b07a29cae31bf9fd5622842f833f25084560533d529e3b36186'
 $checksumType = 'sha256'
 
 $packageArgs = @{